chore(Chip): separate Chip and ChipGroup docs
