+++ packages/react-core/src/components/Chip/Chip.tsx
