+++ packages/react-core/src/components/ChipGroup/index.ts
@@ -1,2 +1,1 @@
 export * from './ChipGroup';
-export * from './Chip';
