+++ packages/react-core/src/components/Select/Select.tsx
@@ -20,7 +20,8 @@ import {
   KeyTypes,
   SelectFooterTabbableItems
 } from './selectConstants';
-import { Chip, ChipGroup, ChipGroupProps } from '../ChipGroup';
+import { ChipGroup, ChipGroupProps } from '../ChipGroup';
+import { Chip } from '../Chip';
 import { Spinner } from '../Spinner';
 import {
   keyHandler,
