+++ packages/react-core/src/components/Toolbar/ToolbarFilter.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { ToolbarItem, ToolbarItemProps } from './ToolbarItem';
-import { ChipGroup, Chip } from '../ChipGroup';
+import { ChipGroup } from '../ChipGroup';
+import { Chip } from '../Chip';
 import { ToolbarContentContext, ToolbarContext } from './ToolbarUtils';
 import { PickOptional } from '../../helpers/typeUtils';
